Compile both lib and dylib.
Dylib is needed for compiler plugins, lib is needed
so the #[macro_use] doesn't add a .so dependencies to the binaries.

It would probably require doing something in build script
to enable dylib only conditionally, so for simplicity
it's always enabled.

Also, it seems that rust ignores #![crate_type = "..."] when
invoked through cargo.